+++ upcloud/testdata/upcloud_loadbalancer/loadbalancer_minimal.tf
@@ -9,13 +9,13 @@ variable "zone" {
 }
 
 resource "upcloud_loadbalancer" "this" {
-  name              = "${var.basename}lb"
-  plan              = "development"
-  zone              = var.zone
+  name = "${var.basename}lb"
+  plan = "development"
+  zone = var.zone
 
   networks {
-    type = "public"
-    name = "public"
+    type   = "public"
+    name   = "public"
     family = "IPv4"
   }
 }