+++ CHANGELOG.md
@@ -5,6 +5,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [5.12.0] - 2024-10-10
+
 ### Added
 
 - upcloud_loadbalancer_frontend_rule: added `matching_condition` field.
@@ -783,7 +785,8 @@ Updated upcloud-go-api, added build/CI scripts, and repackaged 0.1.0 as 1.0.0.
 - resource_upcloud_firewall_rule removed and replaced by resource_upcloud_firewall_rules
 - resource_upcloud_zone removed and replaced by zone and zones datasources
 
-[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.11.3...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.12.0...HEAD
+[5.12.0]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.11.3...v5.12.0
 [5.11.3]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.11.2...v5.11.3
 [5.11.2]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.11.1...v5.11.2
 [5.11.1]: https://github.com/UpCloudLtd/terraform-provider-upcloud/compare/v5.11.0...v5.11.1